+++ src/main/java/miyucomics/hexcassettes/CassetteCastEnv.kt
@@ -9,9 +9,8 @@ import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.util.Hand
 
 class CassetteCastEnv(caster: ServerPlayerEntity, castingHand: Hand) : PlayerBasedCastEnv(caster, castingHand) {
-	override fun produceParticles(particles: ParticleSpray, pigment: FrozenPigment) {}
-
 	override fun getCastingHand(): Hand = this.castingHand
+	override fun produceParticles(particles: ParticleSpray, pigment: FrozenPigment) {}
 	override fun getPigment(): FrozenPigment = IXplatAbstractions.INSTANCE.getPigment(this.caster)
 
 	public override fun extractMediaEnvironment(costLeft: Long, simulate: Boolean): Long {