+++ src/main/java/miyucomics/hexcassettes/patterns/OpFree.kt
@@ -3,19 +3,19 @@ package miyucomics.hexcassettes.patterns
 import at.petrak.hexcasting.api.casting.asActionResult
 import at.petrak.hexcasting.api.casting.castables.ConstMediaAction
 import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
+import at.petrak.hexcasting.api.casting.eval.env.PlayerBasedCastEnv
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.iota.NullIota
+import at.petrak.hexcasting.api.casting.mishaps.MishapBadCaster
 import miyucomics.hexcassettes.HexcassettesAPI
 import net.minecraft.server.network.ServerPlayerEntity
 
 class OpFree : ConstMediaAction {
 	override val argc = 0
 	override fun execute(args: List<Iota>, env: CastingEnvironment): List<Iota> {
-		val caster = env.castingEntity
-		if (caster != null && caster is ServerPlayerEntity) {
-			val playerState = HexcassettesAPI.getPlayerState(caster)
-			return (playerState.ownedCassettes - playerState.queuedHexes.size).asActionResult
-		}
-		return listOf(NullIota())
+		if (env !is PlayerBasedCastEnv)
+			throw MishapBadCaster()
+		val playerState = HexcassettesAPI.getPlayerState((env.castingEntity as ServerPlayerEntity))
+		return (playerState.ownedCassettes - playerState.queuedHexes.size).asActionResult
 	}
 }